Refactor to remove redundant code Part of #729
